(WIP) band folding
dear heavens, the output of this is terrible!

Looks like there was an unnoticed bug in the code which generates
the input positions.json.  Although it was fixed to always produce
supercells with 120 degree cell angles, the individual layers still
have 60 degree unit cells.

That code isn't even part of this repository...!