a miserably large commit I can't possibly document
but I'll try anyways:

did some more refactoring of functionality formerly in the uncrosser
and the oracle; this resulted in a lot of things getting shoved around
in the Shakefile as more responsibilities were pushed into it.

EigenvectorCache is a new horrible thing torn from the heart of
the uncrosser, but its purpose is far more limited in scope so
hopefully I won't need to touch it again for a while.

then I tried to fix band folding

I wasted my entire time at Penn state trying to get a nice, working,
and thoroughly-tested Hnf implementation to fix one bug in the
folding, until finally realizing that all I really needed was the
number of independent k point images along each axis, for which
a brute force algorithm can be easily chucked out in 20 minutes

um

then I added save-stating to the shakefile
(for faster debugging of broken stuff like band unfolding)
(which is still horribly broken)