Filtered-Folded plots and zip adapters
in which a surprising amount of code is written to accomplish the
seemingly simple task of putting filtered and folded bands on the
same plot.

Filtering has been fixed so that gnuplot does not attempt to draw
lines across the discontinuities it introduces. To get the subtle
details correct I had to drop the text-based filtering strategy
and work with an actual data structure.  This made for a bit of trouble.

ashamedly the core issues are my own undoing; the haskell code has
primarily worked with an alternate form of the gnuplot data files
recasted into JSON, but I required a form closer to the original.
This is described in a bit more detail in a comment in the .post/bandplot
section of Shakefile-inner.hs.

---

Zipped operations on nested containers in Haskell have always been
a tough area for me, so I made an Applicative adapter for MonadZip
to help cope with this.  The way it is ultimately used in this
commit isn't... quite what I initially had in mind when I wrote it.
But it found a use nonetheless.

(Truth be told, I am honestly not sure what the sequenceA I wrote
 in transpose' expands to in terms of Vector functions)