add this to version control I guess
positions.py currently shells out to this rust script which
implements brute force HNF searching for the umpteen-billionth
time, so I guess we should check it in.

It was never supposed to see the light of day, but then it did anyway.
(read the comment that prefaces src/main.rs)

Now the biggest timesink in positions.py seems to be
MoirePattern.commensurate_cell() which is--guess what--another
implementation of more or less the same thing.
Tsk.