what was learned from NOT converting the plot code
I tried to convert the plotting code to the new system, but the
results could only be described as "abject failure".
The plot code is simply too ad-hoc in nature to fit into
such a disciplined scheme.

This commit just contains a couple of inoffensive edits that I figured
were worth keeping as I tossed the rest out.

In hindsight, the plot code is fine just the way it is; there's no "process"
to factor out, and in fact the code in its current state already serves to
summarize a great deal of subtle variations that exist between a large number
of related outputs.

In light of this I've begun to form a `main` ruleset, where I'm now collecting
the messier, unstabler rules that help define and link together the various
instances of the cleaner components that were factored out.

Always keep your dirty laundry separate!