fix bug with z positions (cause of foldling issue)
a fencepost error slipped by in my edit to make-poscar
causing z positions to be offset by a constant amount.

Surprisingly, this has a non-trivial impact on the computation,
and fixing it resolved the discrepancy in the folded bands.
Since that makes no physical sense, I still need to look more into
this...