add "touch-mode" rules
I found myself in need of the ability to manually modify a file
and see how it affected dependencies in the build tree

while shake has some functionality for marking files as "clean"
(and even more upcoming in 0.16), it appears that all of this
functionality is explicitly designed to prevent this sort of usage.

Hence these new "keep" and "keep-all" rules.