Local rules
bang-for-the-buck feature (the buck was ~ 1 hour, incl. debugging)

yesterday I had what seemed to be an awesome idea; what if I could put
shake rules files in each directory and load them on demand from my main
script?

Well, we can't quite do that, since Haskell is a compiled language.
But we can at least implement a DSL for providing some subset
  of Shakefile functionality.

This gives the power to specify dependencies of the file-copying
variety (the most common) in a declarative style without necessarily
having to modify the one-and-only main Shakefile